Change ISO 15924 codes to be title-case (e.g. Latn instead of latn), as the standard uses title-case script codes and the data set previously used all-lowercase codes.
Also sort the ISO 15924 codes for each entry alphabetically, as there is no
semantic meaning we intended to convey by the ordering. For example, some
language varieties listed latn,cyrl and others listed cyrl,latn -- but this did not mean that we
thought the Latin script is more common than the Cyrillic script for the first
case, and vice versa for the second case. To avoid confusion, just sort the ISO
15924 codes alphabetically for each language variety listed.

PiperOrigin-RevId: 460443522